requirements.txt
flask==3.0.0
flask-cors==4.0.0
redis==5.0.1
pyyaml==6.0.1
pydantic==2.5.0
